requirements.txt
Flask==2.2.5
flask-cors==3.0.10
pymongo==4.4.0
scikit-learn==1.5.2
nltk==3.8.1
joblib==1.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
werkzeug==2.2.3
requests==2.31.0

cryptography==41.0.3